+++ nextjs-app/app/components/TestimonialBlock.module.css
@@ -1,0 +1,4 @@
+.testimonialGrid {
+  grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
+  max-width: calc(3 * 350px + 2 * 2rem);
+}
